requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone-client
boto3
python-dotenv
PyMuPDF